src/model/requirements.txt
torch==2.1.0
transformers==4.40.0
diffusers==0.27.2
accelerate==0.27.2
safetensors==0.4.2
Pillow
httpx
fastapi
uvicorn